apps/users/serializers.py
from django.shortcuts import render
from rest_framework import routers, serializers, viewsets
from django.contrib.auth.models import User
# Create your views here.
class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ['username', 'email', 'is_staff']
