requirements.txt
absl-py==2.1.0
adrf==0.1.8
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asgiref==3.8.1
astunparse==1.6.3
async-property==0.2.2
attrs==24.2.0
beautifulsoup4==4.12.3
bertopic==0.16.4
black==24.10.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
decorator==4.4.2
deepface==0.0.93
Django==5.1.3
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
drf-yasg==1.21.8
edge-tts==6.1.15
filelock==3.16.1
fire==0.7.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.67.1
gTTS==2.5.3
gunicorn==23.0.0
h5py==3.12.1
hdbscan==0.8.39
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.6.0
libclang==18.1.1
llvmlite==0.43.0
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
moviepy==1.0.3
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pexels-api==1.0.1
pi==0.1.2
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
proglog==0.1.10
propcache==0.2.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pydub==0.25.1
Pygments==2.18.0
pynndescent==0.5.13
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
retina-face==0.0.17
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.66.6
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.2
wrapt==1.16.0
yarl==1.17.1
